pyproject.toml
[project]
name = "flux-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "auth0-fastapi-api>=1.0.0b5",
    "dotenv>=0.9.9",
    "fastapi>=0.115.14",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pygithub>=2.8.1",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.34.3",
]

[tool.pyright]
typeCheckingMode = "strict"
